moved test from the server package to tools directory
