+++ go/server/git_test.go
@@ -23,8 +23,6 @@ import (
 	"os"
 	"testing"
 
-	"github.com/vitessio/arewefastyet/go/tools/git"
-
 	qt "github.com/frankban/quicktest"
 )
 
@@ -100,18 +98,4 @@ func TestPullLocalVitess(t *testing.T) {
 	qt.Assert(t, err, qt.IsNil)
 	err = s.pullLocalVitess()
 	qt.Assert(t, err, qt.IsNil)
-}
-
-func TestGetCommitHash(t *testing.T) {
-	// Create a temporary folder and try setup vitess
-	tmpDir, err := ioutil.TempDir("", "setup_vitess_*")
-	defer os.RemoveAll(tmpDir)
-	s := Server{
-		localVitessPath: tmpDir,
-	}
-	err = s.setupLocalVitess()
-	qt.Assert(t, err, qt.IsNil)
-	out, err := git.GetCommitHash(s.getVitessPath())
-	qt.Assert(t, err, qt.IsNil)
-	qt.Assert(t, len(out), qt.Equals, 40)
 }
